pyproject.toml
[tool.poetry]
name = "jira"
version = "0.1.0"
description = ""
authors = ["Miguel Henriques <github@jmh.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
